+++ app/src/main/java/gateway/soap/request/ReqFileDelete.java
@@ -1,8 +1,9 @@
 package gateway.soap.request;
 
+import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
 
 public class ReqFileDelete extends Authorization
 {
-	public UUID[] fileUUID; // 1+
+	@NotNull public UUID[] fileUUID; // 1+
 }