+++ app/src/main/java/gateway/soap/request/ReqFileNewDir.java
@@ -1,9 +1,10 @@
 package gateway.soap.request;
 
+import jakarta.validation.constraints.NotEmpty;
 import java.util.UUID;
 
 public class ReqFileNewDir extends Authorization
 {
-	public String directoryName;
+	@NotEmpty public String directoryName;
 	public UUID location;
 }