+++ app/src/test/java/gateway/ITServiceAuth.java
@@ -19,6 +19,31 @@ class ITServiceAuth
 {
 	@BeforeEach void setup () { Config.initializeFromEnv (); }
 
+	@Test void Register ()
+	{
+
+		Credentials credentials = new Credentials (UUID.randomUUID ().toString (), "pass");
+
+		// 201
+		ResSession res = CtrlAccountRegister.account_register (credentials);
+		assertEquals (201, res.code, "Register successfully");
+
+		// 500 - user already exists
+		res = CtrlAccountRegister.account_register (credentials);
+		assertEquals (500, res.code, "Internal server error");
+
+		// 500 - without some fields
+		credentials.username = null;
+		res = CtrlAccountRegister.account_register (credentials);
+		assertEquals (400, res.code, "Field: must not be null");
+
+		// 500 - all fails
+		TestUtilConfig.makeInvalidAll ();
+		credentials.username = UUID.randomUUID ().toString ();
+		res = CtrlAccountRegister.account_register (credentials);
+		assertEquals (500, res.code, "Internal error");
+	}
+
 	@Test void Authenticate ()
 	{
 		// 200